backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
openai
pydantic
python-multipart
pypdf
groq
beautifulsoup4
duckduckgo-search
requests
dnspython
